designed and fixed a lot of stuff

+++ WindowsFormsApp1/Code/Choose.cs
@@ -17,7 +17,7 @@ namespace WindowsFormsApp1
         public static string note;
         public static int price;
         public static int discount;
-        public static double debt;
+        public static int debt;
         public static string worker = "אורח";
         // public static string location;
     }
